Extract ticket command handlers from flow_commands into dedicated ticket_commands module
